// src/client/src/shared/const/enums.ts
export enum INPUT_TYPE {
    USERNAME = 0,
    EMAIL = 1,
    OLD_PASSWORD = 2,
    PASSWORD = 3,
    CONFRIM_PASSWORD = 4,
    ENTITY_NAME = 5,
    ENTITY_DESCRIPTION = 6,
    THREAD_NAME = 7,
    THREAD_DESCRIPTION = 8,
    NON_PLAYER_CHARACTER_NAME = 9,
    NON_PLAYER_CHARACTER_DESCRIPTION = 10
}

export enum NON_PLAYER_CHARACTER_FILTER {
    ALL = 0,
    AVAILABLE_FOR_ADVENTURES = 1,
    AVAILABLE_FOR_PLAYER_CHARACTERS = 2
}

export enum TABLE_DISPLAY_MODE {
    TO_SHOW = 0,
    TO_ADD_TO = 1
}

export enum SceneType {
    NORMAL = 0,
    ALTERED = 1,
    INTERRUPTED = 2
}